requirements.txt
pytest==7.2.1
Appium-Python-Client==3.1.1